Chimera build for Linux can play the game for the most part; fixed a bunch of regressions

+++ df2_reimpl/src/Primitives/rdFace.h
@@ -17,7 +17,7 @@ typedef struct rdVertexIdxInfo
     rdVector3* vertices;
     rdVector2* extraUV;
     float* field_14;
-    int* field_18;
+    float* field_18;
 } rdVertexIdxInfo;
 
 typedef struct rdFace
